// projects/dashboard/src/app/interfaces/oanda/oanda.ts
export type OrderStateFilter =
  | 'PENDING'
  | 'FILLED'
  | 'TRIGGERED'
  | 'CANCELLED'
  | 'ALL';
// ...
export interface Order {
// ...
  instrument: string;
// ...
}

/** Get a list of Orders for an Account */
export interface Orders {
  /**The list of Order detail objects */
  orders: Order[];

  /**
   * The ID of the most recent Transaction created for the Account.
   */
  lastTransactionID: string;
}